+++ iCookTV/Helpers/Tracker.swift
@@ -24,30 +24,17 @@
 //  SOFTWARE.
 //
 
+import FBSDKTVOSKit
+import Firebase
 import Foundation
-import Crashlytics
-import Fabric
-import TreasureData_iOS_SDK
 
 enum Tracker {
-
-  static let defaultDatabase = "icook_tvos"
-  static let sessionsTable = "sessions"
-
   static func setUpAnalytics() {
+    Settings.isAutoInitEnabled = false
     #if TRACKING
-      Crashlytics.start(withAPIKey: iCookTVKeys.CrashlyticsAPIKey)
-      Fabric.with([Crashlytics.self])
-
-      TreasureData.initializeApiEndpoint("https://in.treasuredata.com")
-      TreasureData.initialize(withApiKey: iCookTVKeys.TreasureDataAPIKey)
-      TreasureData.sharedInstance().enableAutoAppendUniqId()
-      TreasureData.sharedInstance().enableAutoAppendModelInformation()
-      TreasureData.sharedInstance().enableAutoAppendAppInformation()
-      TreasureData.sharedInstance().enableAutoAppendLocaleInformation()
-
-      TreasureData.sharedInstance().defaultDatabase = Tracker.defaultDatabase
-      TreasureData.sharedInstance().startSession(Tracker.sessionsTable)
+      Settings.appID = iCookTVKeys.FacebookAppID
+      ApplicationDelegate.initializeSDK(nil)
+      FirebaseApp.configure()
     #endif
   }
 
@@ -55,11 +42,7 @@ enum Tracker {
     DispatchQueue.global().async {
       Debug.print(pageView)
       #if TRACKING
-        Answers.logCustomEvent(withName: pageView.name, customAttributes: pageView.details)
-        TreasureData.sharedInstance().addEvent(pageView.attributes, database: defaultDatabase, table: "screens")
-      #endif
-      #if TRACKING && DEBUG
-        TreasureData.sharedInstance().uploadEvents()
+
       #endif
     }
   }
@@ -68,11 +51,7 @@ enum Tracker {
     DispatchQueue.global().async {
       Debug.print(event)
       #if TRACKING
-        Answers.logCustomEvent(withName: event.name, customAttributes: event.details)
-        TreasureData.sharedInstance().addEvent(event.attributes, database: defaultDatabase, table: "events")
-      #endif
-      #if TRACKING && DEBUG
-        TreasureData.sharedInstance().uploadEvents()
+
       #endif
     }
   }
@@ -87,17 +66,8 @@ enum Tracker {
       Debug.print(description, file: file, function: function, line: line)
 
       #if TRACKING
-        Answers.logCustomEvent(withName: "Error", customAttributes: ["Description": description])
-        TreasureData.sharedInstance().addEvent([
-          "description": description,
-          "function": "\(file.typeName).\(function)",
-          "line": line
-        ], database: defaultDatabase, table: "errors")
-      #endif
-      #if TRACKING && DEBUG
-        TreasureData.sharedInstance().uploadEvents()
+
       #endif
     }
   }
-
 }