+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Replace Quick and Nimble with XCTest [#12](https://github.com/polydice/iCook-tvOS/pull/12)
 * Swift 5 [#13](https://github.com/polydice/iCook-tvOS/pull/13)
 * Update gems and migrate to GitHub Actions [#14](https://github.com/polydice/iCook-tvOS/pull/14)
+* Clean up tracking [#15](https://github.com/polydice/iCook-tvOS/pull/15)
 
 ## v1.0.0